Add stub data for creating project correctly.

+++ src/main/java/com/softjourn/coin/server/dto/FoundationProjectDTO.java
@@ -16,7 +16,7 @@ public class FoundationProjectDTO {
   private String adminId;
   private String creatorId;
   private Integer fundingGoal;
-  private Integer deadline;
+  private Long deadline;
   private boolean closeOnGoalReached;
   private boolean withdrawAllowed;
   private String mainCurrency;
